[~] Sync with changed signal recorder..
Signal recorder now stores signals as durations, not as timestamps.

This simplified our code a lot.

But we feel we observe that stored signals are stored a bit more
longer than real signals (like 384 us instead of 334 us).

This may be due fact that storing signal now requires subtraction
of time record. Which is not light operation on scale of CPU cycles.

Anyway, my both IR appliances that I use for testing work with this
version.