[^] Restored UART 115200 speed..
We can process input at 115200 as long as we're not recording.

If we are recording, one more counter is running and overhead for
it's advancing is the drop that overflows UART buffer before parser
can suck it.

We won't add any software blocking for this. Because this situation
can disappear with change of modules. Faster parser or buffered UART
or increased period of recorder's timer or faster processing for
this/other event handlers..